Migrate request validator.
