+++ oauthlib/oauth2/rfc6749/grant_types/base.py
@@ -10,9 +10,9 @@ from itertools import chain
 
 from oauthlib.common import add_params_to_uri
 from oauthlib.oauth2.common import errors
+from oauthlib.oauth2.common.request_validator import RequestValidator
 from oauthlib.oauth2.rfc6749 import utils
 
-from ..request_validator import RequestValidator
 
 log = logging.getLogger(__name__)
