+++ oauthlib/oauth2/rfc6749/grant_types/refresh_token.py
@@ -9,9 +9,9 @@ import json
 import logging
 
 from oauthlib.oauth2.common import errors
+from oauthlib.oauth2.common.request_validator import RequestValidator
 
 from .. import utils
-from ..request_validator import RequestValidator
 from .base import GrantTypeBase
 
 log = logging.getLogger(__name__)
